redux: organize product slices
