+++ ecommerce/src/components/AddProductForm.tsx
@@ -1,7 +1,7 @@
 import React, { FunctionComponent, useState } from 'react';
 import { useDispatch } from 'react-redux';
-import { addProduct } from '../redux/reducers/allProductsSlice';
-import { updateIsAddingNewProductState } from '../redux/reducers/isAddingNewProductSlice';
+import { addProduct } from '../redux/reducers/product/allProductsSlice';
+import { updateIsAddingNewProductState } from '../redux/reducers/product/isAddingNewProductSlice';
 import { addNewProduct } from '../services/product/addNewProduct.service';
 import { ProductType } from '../types/ProductType';
