+++ ecommerce/src/redux/reducers/product/allProductsSlice.ts
@@ -1,5 +1,5 @@
 import { createSlice } from '@reduxjs/toolkit';
-import { ProductType } from '../../types/ProductType';
+import { ProductType } from '../../../types/ProductType';
 
 let initialState: ProductType[] = [];
